+++ redditrepostsleuth/core/celery/app.py
@@ -1,4 +1,7 @@
 import os
+
+from celery.signals import after_setup_logger
+
 print(os.getcwd())
 # Celery is broken on windows
 import sys
@@ -11,6 +14,8 @@ registry.enable('pickle')
 celery = Celery('tasks')
 celery.config_from_object('redditrepostsleuth.core.celery.celeryconfig')
 
-
+@after_setup_logger.connect
+def setup_loggers(logger, *args, **kwargs):
+    logger.handlers = []
 if __name__ == '__main__':
     celery.start()