+++ utility_scripts/push_shift_backfill_beta.py
@@ -13,6 +13,9 @@ from redditrepostsleuth.core.db.uow.sqlalchemyunitofworkmanager import SqlAlchem
 from redditrepostsleuth.core.logging import log
 
 
+
+oldest_timestamp = None
+
 config = Config()
 uowm = SqlAlchemyUnitOfWorkManager(get_db_engine(config))
 
@@ -38,29 +41,33 @@ def fetch_results(url: Text):
 
     return response['payload']
 
-def get_from_api(oldest_created = None):
-    base_url = 'https://api.pushshift.io/reddit/search/submission?size=2000&sort_type=created_utc&sort=desc'
-    if oldest_created:
-        url = base_url + '&before=' + str(oldest_created)
+def get_from_beta_api(oldest_sid = None):
+    base_url = 'https://beta.pushshift.io/search/reddit/submissions?size=1000&sort_type=created_utc&sort=desc'
+    if oldest_sid:
+        url = base_url + '&max_sid=' + str(oldest_sid)
     else:
         url = base_url
     return fetch_results(url)
 
 def parse_and_submit_to_queue(data) -> Text:
-    log.debug('Oldest: %s | Newest: %s', datetime.utcfromtimestamp(data[-1]['created_utc']),
-              datetime.utcfromtimestamp(data[0]['created_utc']))
+    log.debug('Oldest: %s | Newest: %s', datetime.utcfromtimestamp(data['data'][-1]['created_utc']),
+              datetime.utcfromtimestamp(data['data'][0]['created_utc']))
 
     try:
-        save_pushshift_results.apply_async((data,), queue='pushshift')
+        save_pushshift_results.apply_async((data['data'],), queue='pushshift')
     except Exception as e:
-        log.exception('Failed to send to celery', exc_info=False)
+        log.exception('Failed to send to pushshift', exc_info=False)
 
+    try:
+        return data['data'][-1]['sid']
+    except ValueError:
+        log.error('Failed to get oldest SID')
 
-oldest_created = None
+oldtest_sid = None
 while True:
-    results = get_from_api(oldest_created)
+    results = get_from_beta_api(oldtest_sid)
     if not results:
         continue
-    oldest_created = results['data'][-1]['created_utc']
+    oldtest_sid = results['data'][-1]['sid']
     parse_and_submit_to_queue(results['data'])