+++ changelog.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0.2] - 1/30/2021
+* [bugfix] - Bot was failing to check posts on monitored subreddit if subreddit was private
+* [enhancement] - If a post is ingested from a monitored subreddit repost check is performed at ingest time
+* [enhancement] - Split search index by year to enable faster rebuilds and reduce resource usage
+
 ## [1.0.1] - 1/30/2021
 * [feature] - Add send_repost_modmail to monitored subs.  This will trigger a mod mail when a repost is detected
 * [bugfix] - Comment on link reposts was pointing to itself as the source